requirements.txt
transformers
torch
Image
python-dotenv
langchain-openai
playsound
